+++ ckb-mercury-sdk/src/main/java/org/nervos/mercury/model/req/payload/DaoDepositPayload.java
@@ -2,13 +2,11 @@ package org.nervos.mercury.model.req.payload;
 
 import org.nervos.mercury.model.req.From;
 
-import java.math.BigInteger;
-
 public class DaoDepositPayload {
   public From from;
   public String to;
   public long amount;
-  public BigInteger feeRate;
+  public Long feeRate;
 
   protected DaoDepositPayload() {
   }
